# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/geofences": {
      "get": {
        "description": "Returns all geofence definitions for the operations dashboard.",
        "operationId": "AdminGeofence_listGeofences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of geofences"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List geofences (supervisor / admin)",
        "tags": ["Admin geofences (supervisor / admin)"]
      },
      "post": {
        "description": "Creates a circular work area. Sign-in is allowed when the device is inside any active geofence.",
        "operationId": "AdminGeofence_createGeofence",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["label", "centerLatitude", "centerLongitude", "radiusMeters"],
                "properties": {
                  "label": {
                    "type": "string",
                    "example": "Nairobi CBD"
                  },
                  "centerLatitude": {
                    "type": "number",
                    "example": -1.286389
                  },
                  "centerLongitude": {
                    "type": "number",
                    "example": 36.817223
                  },
                  "radiusMeters": {
                    "type": "number",
                    "example": 5000,
                    "description": "Radius from center in meters"
                  },
                  "isActive": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Geofence created"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create geofence (supervisor / admin)",
        "tags": ["Admin geofences (supervisor / admin)"]
      }
    },
    "/admin/geofences/{id}": {
      "patch": {
        "description": "Partial update (label, center, radius, isActive).",
        "operationId": "AdminGeofence_updateGeofence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Geofence id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "centerLatitude": {
                    "type": "number"
                  },
                  "centerLongitude": {
                    "type": "number"
                  },
                  "radiusMeters": {
                    "type": "number"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Geofence updated"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          },
          "404": {
            "description": "Geofence id not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update geofence (supervisor / admin)",
        "tags": ["Admin geofences (supervisor / admin)"]
      }
    },
...
    "/me": {
      "get": {
        "description": "Returns the authenticated user's profile details.",
        "operationId": "Me_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cmad4p0bo0000iib0i0l9e8wk",
                    "fullName": "Jamal Salim",
                    "phone": "+254712345678",
                    "role": "promoter",
                    "gender": "male",
                    "regionId": "nairobi-west"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token"
          },
          "404": {
            "description": "User profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": ["Me"]
      },
      "patch": {
        "description": "Updates one or more profile fields for the authenticated user.",
        "operationId": "Me_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              },
              "examples": {
                "updateNameAndRegion": {
                  "summary": "Update name and region",
                  "value": {
                    "fullName": "Jamal S. Salim",
                    "regionId": "nairobi-east"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cmad4p0bo0000iib0i0l9e8wk",
                    "fullName": "Jamal S. Salim",
                    "phone": "+254712345678",
                    "role": "promoter",
                    "gender": "male",
                    "regionId": "nairobi-east"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": ["Me"]
      }
    },
    "/me/location/history": {
      "get": {
        "description": "Returns recent location pings for the authenticated user, newest first.",
        "operationId": "Me_listLocationHistory",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum rows to return (1–100, default 50)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location ping rows",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "cmad4p0bo0000iib0i0l9e8wk",
                      "latitude": -1.286389,
                      "longitude": 36.817223,
                      "recordedAt": "2026-05-08T18:20:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List own location check-ins",
        "tags": ["Me"]
      }
    },
    "/me/location": {
      "post": {
        "description": "Stores a location ping for the authenticated user.",
        "operationId": "Me_updateMeLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              },
              "examples": {
                "nairobiDowntown": {
                  "summary": "Nairobi CBD location ping",
                  "value": {
                    "latitude": -1.286389,
                    "longitude": 36.817223
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recorded location ping",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "userId": "cmad4p0bo0000iib0i0l9e8wk",
                    "latitude": -1.286389,
                    "longitude": 36.817223,
                    "recordedAt": "2026-05-08T18:20:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record user location",
        "tags": ["Me"]
      }
    }
  },
  "info": {
    "title": "Engaged Sales API",
    "description": "API documentation for Engaged Sales backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Jamal Salim"
          },
          "gender": {
            "type": "string",
            "example": "male",
            "enum": ["male", "female", "other"]
          },
          "regionId": {
            "type": "string",
            "example": "nairobi-west"
          }
        }
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": -1.286389,
            "description": "Latitude in decimal degrees"
          },
          "longitude": {
            "type": "number",
            "example": 36.817223,
            "description": "Longitude in decimal degrees"
          }
        },
        "required": ["latitude", "longitude"]
      }
    }
  }
}